--- wntr/tests/networks_for_testing/Anytown.inp
@@ -1,205 +1,0 @@
-[TITLE]
-ANYTOWN example
-
-[JUNCTIONS]
-;---------------------------------------
-;	Elev	Demand
-; ID	ft	 gpm		Pattern
-;---------------------------------------
-  1	20	500		1
-  2	50	200		1
-  3	50	200		1
-  4	50	600		1
-  5	80	600		1
-  6	80	600		1
-  7	80	600		1
-  8	80	400		1
-  9    120	400		1
-  10   120	400		1
-  11   120	400		1
-  12    50	500		1
-  13    50	500		1
-  14	50	500		1
-  15	50	500		1
-  16   120	400		1
-  17   120     1000		1
-  18    50	500		1
-  19	50     1000		1
-  20	20        0		1
-  21    50        0		1
-  22   120        0		1
-
-
-[RESERVOIRS]
-;ID              	Head        	Pattern         
- 40	 		10                      	;
-
-
-[TANKS]
-;-------------------------------------------
-;	Elev	Init	 Min	 Max	Diam
-; ID	(ft)	Level	Level	Level	(ft)
-;-------------------------------------------
-  41	215	10	10	35	32.65  0  ;
-  42    215	10	10	35	32.65  0  ;
-
-
-[PIPES]
-;--------------------------------------------------------
-;	Start	End	Length	Diam	Rough	Location
-; ID	Node	Node	 (ft)	(in)	Coef.
-;--------------------------------------------------------
-  1      1	 2	12000	 12	120     0	open	;Res
-  2	 1      12	12000	 12	 70	0	open	;city
-  3      1	13	12000	 16	 70	0	open	;city
-  4      1	20	  100	 30	130	0	open	;city
-  5	 2 	 3	 6000	 10	120     0	open	;Res
-  6      2	 4 	 9000    10	120     0	open	;Res
-  7      2	13	 9000	 12	 70     0	open	;Res
-  8      2	14	 6000	 10	120     0	open	;Res
-  9	 3	 4	 6000	 10	120     0	open	;Res
-  11     4	 8	12000	  8	120     0	open	;Res
-  12	 4	15	 6000	 10	120     0	open	;Res
-  17     8	 9	12000	  8	120     0	open	;Res
-  18     8	15	 6000	 10	120     0	open	;Res
-  19     8	16	 6000	  8	120     0	open	;Res
-  20     8	17       6000	  8	120     0	open	;Res
-  21     9	10	 6000	  8	120     0	open	;Res
-  22    10	11	 6000	  8	120     0	open	;Res
-  23    10	17	 6000	 10	120     0	open	;Res
-  24    11	12	 6000	  8	120     0	open	;Res
-  26    12	17	 6000	 10	120     0	open	;Res
-  27    12	18       6000	  8	 70	0	open	;city
-  28    13	14	 6000	 12	 70	0	open	;city
-  29    13	18       6000	 12	 70	0	open	;city
-  30    13	19	 6000	 10	 70	0	open	;city
-  31    14	15	 6000	 12	 70	0	open	;city
-  32    14	19	 6000	 10	 70	0	open	;city
-  33    14	21	  100	 12	120	0	open	;city
-  34    15	16	 6000	 10	 70	0	open	;city
-  35    15	19	 6000	 10	 70	0	open	;city
-  36    16	17	 6000	  8	120	0	open	;Res
-  37    16	18	 6000	 12	 70	0	open	;city
-  38    16	19	 6000	 10	 70	0	open	;city
-  39    17	18	 6000	  8	120	0	open	;Res
-  40    17	22	  100	 12	120	0	open	;Res
-  41    18	19	 6000	 10	 70	0	open	;city
-  142	 21	41	    1	 12	120	0	open	;city
-  143	 22	42	    1	 12	120	0	open	;Res
-  110	 4	 5	 6000	 0.0001	130     0	open	;New
-  113    5	 6	 6000	 0.0001	130     0	open	;New
-  114    6	 7       6000	 0.0001	130     0	open	;New
-  115    6	 8       6000	 0.0001	130     0	open	;New
-  116    7	 8       6000	 0.0001	130     0	open	;New
-  125    11	17	 9000	 0.0001	130     0	open	;New
-
-
-[PUMPS]
-;-------------------------------------------------------------------
-;	Start	End   	
-; ID	Node	Node	Parameters		Energy Price
-;-------------------------------------------------------------------
-  78	40	20      HEAD 2	PATTERN 2	0.12
-  79	40	20      HEAD 2	PATTERN 3	0.12
-  80	40	20      HEAD 2	PATTERN 4	0.12
-
-[ENERGY]
- PUMP		78  	EFFIC E1
- PUMP		79  	EFFIC E1 
- PUMP		80  	EFFIC E1 
-
-
-[PATTERNS]
-;ID              	Multipliers
-;Daily water use pattern
- 1               1.0	1.0	1.0	0.9	0.9	0.9
- 1               0.7	0.7	0.7	0.6 	0.6	0.6         
- 1               1.2   	1.2	1.2	1.3	1.3	1.3         
- 1               1.2	1.2	1.2	1.1	1.1	1.1	           
-            
-;Daily Operating Pump
- 2               0	0           0           0           0           0           
- 2               0	0           0           0           0           0           
- 2               0	0           0           0           0           0           
- 2               0	0           0           0           0           0 	          
-;Daily Alternate Pump
- 3               0	0           0           0           0           0           
- 3               0	0           0           0           0           0           
- 3               0	0           0           0           0           0           
- 3               0	0           0           0           0           0	            
-;Daily Alternate Pump
- 4               0	0           0           0           0           0           
- 4               0	0           0           0           0           0           
- 4               0	0           0           0           0           0           
- 4               0	0           0           0           0           0		
-
-[CURVES]
-;ID              	X-Value     	Y-Value
- 2               	0.          	300        
- 2               	2000      	292       
- 2               	4000      	270       
- 2               	6000      	230       
- 2               	8000      	181      
-
-;ID			flow		efficiency
- E1               	0.          	0        
- E1               	2000      	50       
- E1               	4000      	65       
- E1               	6000      	55       
- E1               	8000      	40  
-
-
-[TIMES]
- Duration           	24:00 
- Hydraulic Timestep 	0:01
- Quality Timestep   	0:01 
- Pattern Timestep   	1:00 
- Pattern Start      	0:00 
- Report Timestep    	1:00 
- Report Start       	0:00 
- Start ClockTime    	18 PM
- Statistic          	NONE
-
-
-
-[OPTIONS]
-;------------------------------------------
- UNITS	    GPM
- QUALITY    NONE
- ACCURACY 0.001
- DEMAND MULTIPLIER 1.0
- Emitter Exponent   	0.5
- Diffusivity        	1.0
- Tolerance          	0.01
-
-
-[COORDINATES]
-;Node            	X-Coord         	Y-Coord
- 1              	7682.33         	3371.15         
- 2              	7633.71         	5737.44         
- 3              	7520.26         	7293.35         
- 4              	6175.04         	7568.88
- 5              	5591.57         	8460.29         
- 14              	6466.77         	5769.85         
- 13              	6450.57         	4424.64         
- 6              	4100.49         	8071.31         
- 15              	5094.45         	6482.09         
- 19              	5332.25         	5332.25         
- 18             	4846.03         	4440.84         
- 12             	4846.03         	3354.94         
- 7  	           	3192.87         	7763.37         
- 11             	3701.63         	2892.06         
- 10             	2317.45         	3588.20         
- 8     	        	3321.19         	6696.55         
- 16             	4017.86         	5615.96         
- 17             	3005.49         	4292.43         
- 9         	    	1305.62         	4850.97         
- 20              	8544.18         	3371.15         
- 40			8700.6			3371.15
- 21              	6094.00         	6499.19         
- 41			6094.00			6600.19
- 22             	2422.68         	4519.09         
- 42 			2422.68			4700.09
-
-
-[END]
